feat: Implement full-screen diff viewer and symlink command
Implemented a full-screen diff viewer triggered by the 'F' key, allowing for
better inspection of file changes. Also, the `link_topsymlinks` command was
integrated, which handles the creation of symlinks for untracked files, editor
configurations, and the creation of a `tmp` directory within a worktree. This
command also attempts to run `direnv allow` if an `.envrc` file is present.

Signed-off-by: Chmouel Boudjnah <chmouel@chmouel.com>